refactor(lexer): created a pseudo namespace for lexer logic
In an effort to organize the code, I've decided to put some functions
behind a pseudo namespace. I am not sure if this is common practice -- I
should look into deno_std to see if they use a similar pattern -- but it
makes sense to me!

In other news, the regex objects are no longer created on every call and
I am in the process of changing the project's formatting.

BREAKING CHANGE: the analyze method is now behind a pseudo namespace